+++ R/subset.R
@@ -36,12 +36,12 @@
     build_tsibble(
       res,
       key = !!new_key, index = !!index(x), index2 = !!index2(x),
-      interval = is_regular(x), ordered = is_ordered(x), validate = FALSE
+      interval = interval(x), ordered = is_ordered(x), validate = FALSE
     )
   } else { # index2 not there
     build_tsibble(
       res,
-      key = !!new_key, index = !!index(x), interval = is_regular(x),
+      key = !!new_key, index = !!index(x), interval = interval(x),
       ordered = is_ordered(x), validate = FALSE
     )
   }